test local slack app
